antialliasing working for shaded

+++ includes/ray_calculate.h
@@ -7,4 +7,4 @@ double scene_intersect_ray(struct object_intersection *closest_intersection,
                            struct scene *scene, struct ray *ray);
 
 struct ray image_cast_ray(const struct rgb_image *image,
-                          const struct scene *scene, size_t x, size_t y);
+                          const struct scene *scene, double x, double y);
